Rewrote parseInt to parseFloat

+++ timeweb/timewebapp/static/js/crud.js
@@ -113,7 +113,7 @@ class Crud {
         }
         that.old_unit_value = singularToLowerCase;
         // +1 to show the line under "Advanced Inputs"
-        $("#fields-wrapper").css("height", $("#advanced-inputs").position().top + $("#advanced-inputs").height() + parseInt($("#advanced-inputs").css("margin-top")) + 1 + $("#fields-wrapper").scrollTop());
+        $("#fields-wrapper").css("height", $("#advanced-inputs").position().top + $("#advanced-inputs").height() + parseFloat($("#advanced-inputs").css("margin-top")) + 1 + $("#fields-wrapper").scrollTop());
     }
     setCrudHandlers() {
         const that = this;
